ProjectCard as component
