+++ index.html
@@ -24,28 +24,28 @@
     <footer>
       <div class="social-media">
         <div class="social-footer">
-          <a href="https://linkedin.com/in/your-twitter" target="_blank" rel="noopener noreferrer">
+          <!--<a href="https://linkedin.com/in/your-twitter" target="_blank" rel="noopener noreferrer">-->
             <div class="footer-icons">
               <ion-icon name="logo-linkedin"></ion-icon>
             </div>
           </a>
         </div>
         <div class="social-footer">
-          <a href="https://youtube.com/your-twitter" target="_blank" rel="noopener noreferrer">
+          <!--<a href="https://youtube.com/your-twitter" target="_blank" rel="noopener noreferrer">-->
             <div class="footer-icons">
               <ion-icon name="logo-youtube"></ion-icon>
             </div>
           </a>
         </div>
-        <div class="social-footer"
-          <a href="https://twitter.com/your-twitter" target="_blank" rel="noopener noreferrer">
+        <div class="social-footer">
+        <!--<a href="https://twitter.com/your-twitter" target="_blank" rel="noopener noreferrer">-->
             <div class="footer-icons">
               <ion-icon name="logo-twitter"></ion-icon>
             </div>
           </a>
         </div>
         <div class="social-footer">
-          <a href="https://wwwinstagram.com/your-twitter" target="_blank" rel="noopener noreferrer">
+          <!--<a href="https://wwwinstagram.com/your-twitter" target="_blank" rel="noopener noreferrer">-->
             <div class="footer-icons">
               <ion-icon name="logo-instagram"></ion-icon>
             </div>
